# syncro/apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Syncro API",
    "description": "Syncro backend REST API.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/sections": {
      "get": {
        "tags": [
          "section-controller"
        ],
        "summary": "List sections",
        "operationId": "listSections",
        "parameters": [
          { "name": "plantId", "in": "query", "required": false, "schema": { "type": "string", "format": "uuid" } },
          {
            "name": "includeInactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sections returned",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SectionListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "section-controller"
        ],
        "summary": "Create a section",
        "operationId": "createSection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Section created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SectionView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sections/{sectionId}": {
      "get": {
        "tags": [
          "section-controller"
        ],
        "summary": "Get a section",
        "operationId": "getSection",
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Section returned",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SectionView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "section-controller"
        ],
        "summary": "Update a section (name, active); deactivation is guarded",
        "operationId": "updateSection",
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Section updated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SectionView"
                }
              }
            }
          },
          "409": {
            "description": "SECTION_HAS_ACTIVE_MACHINE_GROUPS",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SectionView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SectionView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "plantId": {
            "type": "string",
            "format": "uuid"
          },
          "plantCode": {
            "type": "string"
          },
          "plantName": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "enum": [
              "MACHINERY",
              "UTILITY",
              "WORKSHOP"
            ]
          },
          "name": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateSectionRequest": {
        "type": "object",
        "required": [
          "plantId",
          "code",
          "name"
        ],
        "properties": {
          "plantId": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "enum": [
              "MACHINERY",
              "UTILITY",
              "WORKSHOP"
            ]
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1
          }
        }
      },
      "UpdateSectionRequest": {
        "type": "object",
        "required": [
          "name",
          "active"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
      "SectionListResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SectionView"
            }
          }
        }
      }
    }
  }
}